requirements.txt
gradio
imbalanced-learn
joblib
kagglehub
matplotlib
numpy
pandas
Pillow
scikit-learn
